// common/src/types/user-type.ts
export enum UserType {
  System = 'system',
  Superadmin = 'superadmin',
  PartnerSuperuser = 'facility:superuser',
  PartnerManager = 'facility:manager',
  PartnerEmployeeAdmin = 'employee:admin',
  Doctor = 'doctor',
  Diabetologist = 'diabetologist',
  PhysicianAssistant = 'physician:assistant',
  Dietician = 'dietician',
  Nutritionist = 'nutritionist',
  Educator = 'educator',
  CustomerSupport = 'customer:support',
  PartnerRosterManager = 'roster:manager',
  Patient = 'patient',
  OCAdmin = 'oc:admin',
  OCRequestManager = 'oc:request:manager',
  OCDeliveryVolunteer = 'oc:delivery:volunteer',
  Employee = 'employee',
  Staff = 'staff'
}